+++ test/public_surface_test.exs
@@ -289,7 +289,7 @@ defmodule PublicSurfaceTest do
   end
 
   setup do
-    Imp.configure(lm: nil, adapter: Imp.Adapter.Chat, retriever: nil)
+    Imp.configure(lm: nil, adapter: Imp.Adapter.Chat)
     # Restore global Imp.Settings to defaults on exit (test isolation). See dee-fqsr.
     on_exit(&Imp.Settings.reset/0)
     :ok